Parser and Tokenizer rework to static buffers. Among other things.
Removed old unimplemented `--define` argument in tcalc_cli

Changed tcalc_exprtree to be contained and backed by a single array. The
caller must allocate a large enough array to hold all tree data: the parsing
functions no longer can grow the maximum capacity for a parse tree. This
limitation, given that a tcalc expression is usually just a few dozen or
(hundred nodes in extreme cases?) isn't that bad of a trade off.

Relatedly, the tcalc cli uses a global statically allocated buffer of a few
kibibytes to hold all token and expression node data for processing.

Change TCALC_ERR_BAD_ALLOC to TCALC_ERR_NOMEM. Done so that cases of running
out of memory in a buffer and bad allocation could share a single error code.

Changed functions to always parse to a `t

Changed parser to work independently of the `tcalc_ctx` struct. If a variable
is followed by an opening parenthesis, it is assumed that that variable is
a function, and is parsed accordingly. Checking for proper arity based on
function conditions or whether the function actually exists is done during
evaluation of the expression tree. This allows us to keep a single tree with
a changing `tcalc_ctx` in the future and know that changes to a `tcalc_ctx`
will not affect what the structure of our tree should be for a given input
string.

Added 255 argument cap to the number of arguments parsed with a function.
Completely open to raising that number.

Removed tcalc_strlcpy fully. strlcpy is not very good, and shouldn't be copied.

Changed token indices to use an explicit 32-bit signed integer instead of a
`size_t` (which would generally be 64-bits). tcalc will never deal with a
single input that is larger than 4 GB, so the extra bits that double the
size of every token are completely unnecessary.

Remove old vardef things. They will get added if variable definitions ever
get added, which I do not forsee. I'd rather just make a new parser or
something instead of trying to backfill statements and control flow into
tcalc.